pyproject.toml
[tool.poetry]
name = "task_tracker"
version = "0.0.1"
description = "Task Tracker"
authors = ["Shostak Serhii <shostakserhii94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "0.29.0"
sqlalchemy = "^2.0.34"
databases = "^0.9.0"
pydantic = "2.7.1"
poetry = "^1.8.3"
lock = "^2018.3.25.2110"
aiosqlite = "^0.20.0"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
